Minor changes to the picamera docs
Primarily tweaked the preview section; the preview is a GPU-driven overlay that doesn't require X running at all so it'll always hide the Python console if you're working on the Pi. As several users have had an issue stopping this, I've added the Ctrl+D trick for those that find blind typing the stop_preview method call difficult.