Typo/grammar fixes, other general edits
Headers now correctly formatted; for the avoidance of doubt, this is the newly-agreed house style for them.

Removed link to Wikipedia as there were multiple issues with the article and it is not immediately clear which section refers to CEA in the context of displays. Added a link to the heatsink on RS Components.