Update raspistill.md
`less` didn't work for me. When trying to use the arrow keys it scrolled to an empty buffer and there was no way to get back. Might be related to the Norwegian keyboard layout, but I've never had an issue with `less` and Norwegian keyboards on other linux distros. Anyways I guess shift + page up and shift + page down should be more robust.