Update README.md
Moved a section that was after the glossary, to before the glossary. Most of the terms referred to in the glossary were in that section after the glossary. The reader falls off the end of the doc onto the glossary and there are all these references to to terms that have not been covered. So its a bit confusing like that. Its a technical not but not that technical. Don't think there is a need to hide at the end of the doc.