Improved Mac OS SD format
The Mac OS SD format instructions left out the critical step of installing the Master Boot Record without which NOOBS fails.  This was likely overlooked before because many cards come pre-formatted with the MS-DOS master boot record and merely need to be erased, but SD cards previosly used for NOOBS or MAC OSX may need to be repartitioned and at that moment it's critical to re-install the MBR or NOOBS will not boot on a Rasberry Pi.