Added additional note on secuirity concerns
Most tutorials (including the official documentation for the Raspbian) use the very simple way of entering a password in a `wpa_supplicant` profile as **plain text**. Depending on the scenario (for example a server) the user might not want to do that. Using `wpa_passphrase` ensures that we use a secret token instead of plain text password in our configuration file.