Change tool used to find SD card name in install docs (#807)
* Change tool used to find SD card name in install docs

Updated guide to use `lsblk` instead of `df -h` to find out the SD card name. Reason: it has cleaner output for this task and works regardless of whether or not partitions on it are mounted (and it also shows their mountpoints), this means the guide can be simplified. To the best of my knowledge it should also be present by default on any common linux distro (tool is provided in `util-linux` package).

Also emphasised warning about using the correct device name while using `dd`

* copy edits