Add notes about distro upgrades
I found this page while looking for instructions on upgrading to stretch.  This fixes the "dist-upgrade" command to simply "upgrade" so as not to mislead the reader into thinking it's a distro upgrade.  This also adds new heading with a link to an example of how to upgrade to a new version of Raspbian.