Removing insecure examples of wpa_passphrase
The command `wpa_passphrase ssid password` should never be used or suggested to any noob, since shell will put the line into the history, from which the password can be easily recovered along with the ssid. wpa_passphrase is not designed to be used in that way, since it will ask for the password if not added as paramter, and that input will not be stored anywhere. The ability to be able to use commands with passwords only designed for scripts they SHOULD NOT BE USED in a terminal (where history of commands being logged).