Workaround lack of support for 3 level lists
For some reason the website cannot render 3 level lists. Github helpfully marks such lines with a red highlight in its built-in editor. In fact, if you look at bootflow.msd in the github editor you will see that it highlights both 3rd level and 6th level indented list items. So there is something weird going on. In any case, work around this by just not making the top level list items part of a list, but rather separate links spaced apart from each other. Too me it actually makes more sense this way, since the previous top level of the list was just a list of pages, which is not a sequence nor tightly-coupled set, if you look at their titles. I've also added an extra bit of text on the end of the list item for USB host boot to clarify that with host mode boot you have 2 options - MSD and net boot, and those are the only options for host boot - it's not separate.