Correct info on which Pi models support MSD boot (#1030)
* Correct info on which Pi models support MSD boot

Update documentation/msd.md with information on which current Pi models support USB mass storage device booting.

* fix table layout

* Update msd.md

* Add 3A+ to list of Pi's that can do MSD boot

Explanation - originally I had thought that the 3A+ could not do MSD boot. This was a mistake. The reason for my mistake was that I had heard that the 3A+ could not do USB device mode boot, and had got mixed up and thought that it was MSD boot that it could not do. Prompted by a comment on #1032, I checked with google and discovered my mistake.