Clarify use of enable_uart and core_freq options (#1038)
* Clarify use of enable_uart and core_freq options

Recently a forum user got confused by this part of the docs as they are currently written. This change rewords it so it is a bit less confusing.

* use 'primary uart' rather than 'linux console'

Since the Linux console may be disabled, and in any case we are concerned with which way round the uarts are configured, not the presence or otherwise of the Linux console. Also explain what both /dev/serial0 and /dev/serial1 are.

* uart -> UART

* add / to dev/serial1

* 400MHz -> VPU turbo frequency

* Change BT to Bluetooth

the 1 instance I missed previously

* copy edits