Add some content to bootmodes/sdcard.md
documentation/hardware/raspberrypi/bootmodes/sdcard.md is intended to document SD card booting as part of the 'boot modes' documentation - it is currently blank. Add a few sentences about SD card booting, along with a link to the 'installation' section of the docs for those looking for basic SD card installation info.