Rewrite :  Use etcher or 1 example using dd
Previous multiple edits have made these instructions convoluted. 

With the availability of Etcher and Apple's System Integrity Protection restrictions and the variations with internal /external SD drive and new/unformattted/previously used cards and with the risk of overwriting the running system

Is it a better idea to just let etcher be THE way?