Update README.md
DS-lite is very common in the region where I live. Dyndns-services for IPv6 already exist, some of them are for free.
IPv6 access is at moment limited to 25% of worlds internet user, but in some countries like us, be, de, fr, in, gr it an alternative already for more than 40% of the users.