Add the inlets project to the list of alternatives
You can use inlets to access your Raspberry Pi remotely, even if
you have a dynamic IP address assigned that changes periodically.
It is a good alternative to ngrok or Argo Tunnel from Cloudflare
which are closed-source products.
It combines a reverse proxy and a websockets tunnel to expose
the local endpoint (running on a RPi) to the public internet.

Inlets is an open source project part of the CNCF Landscape.

Signed-off-by: Utsav Anand <utsavanand2@gmail.com>