Fix example of default location for ssh keys
The ssh keys are generated on the computer that is used to connect to the Raspberry Pi. The default location for the ssh keys most likely isn't '/home/pi/.ssh/id_rsa', but generally is '~/.ssh/id_rsa'. 
This was confusing, since it could be interpreted as having to generate the ssh keys on the Raspberry Pi itself.