linux/usage/root.md - remove reference to visudo
Adding a user to the 'sudo' group is preferrably, since it does not involve editing any files directory. Also, that's how it is done for the Pi user. The file /etc/sudoers.d/010_pi-nopasswd is only there so the pi user is not prompted for a password when they attempt to sudo.