WIP: Use visudo to edit any sudoers file (#1550)
* WIP: Use visudo to edit any sudoers file

See #1107. 

There are currently 3 pages in this repo which have problems to do with the unsafe editing of sudo-related files:

- /linux/usage/users.md
- /linux/usage/root.md
- /configuration/security.md

There is also /usage/terminal/README.md which contains slightly wrong information (doesn't mention Raspbian has no sudo password on the pi user by default), and a gramatical error (comma splice).

This PR fixes them.

* Update users.md

* Update root.md

* usage/terminal/README.md fixup

* Update security.md

* configuration/security.md - explain visudo syntax check

* linux/usage/root.md - remove reference to visudo

Adding a user to the 'sudo' group is preferrably, since it does not involve editing any files directory. Also, that's how it is done for the Pi user. The file /etc/sudoers.d/010_pi-nopasswd is only there so the pi user is not prompted for a password when they attempt to sudo.

* configuration/security.md - add strong warning about sudoers breakage

* configuration/security.md - make shell prompt an example

* linux/usage/users.md - copy explanation of visudo syntax check from configuration/security.md

* fix users.md

* Update security.md

* Update security.md

* Update users.md