Update linux.md
Added sync as part of the dd write commands. Logically, sync should be executed at the end of the dd writes to the SD card. Organizationally, this page has the sync separated from the dd command by at least half a page, separated by a section on verification, and formatted as text, rather than as the essential command that it is.