Explicitly call-out Vendor Class identifier for PXE
Vendor Class identifier was buried in some tcpdump parsed output further down in the page. Explicitly calling it out up-top, in some environments this is used as an initial pass filter by the DHCP server to determine which set of responses / behavior to use when interacting with the PXE Client system.