Additional / better example for supplicant file.
Just an edit of the headless setup to provide a more elaborate - working wpa_supplicant.conf file for most typical consumer wpa2 personal networks hidden ssid or not. 

This config has been proven and tested to work personally and has always been referred to in our discord community with thousands of users who love Raspberry Pi when the basic template provided on the site doesn't cut it or work for those not willing to learn proper usage of wpa_supplicant.conf entries.