Installation: deprecate NOOBS, tweak page links
Just noticed this page is still recommending NOOBS, so this change aims to update it to point folk to Raspberry Pi Imager instead. I've also made some attempt to make use of the new Foundation-provided content, which somewhat supercedes what is in the installation section of the docs.

The new Foundation-provided page at https://www.raspberrypi.org/software/ is simpler anything in this repo, so I've linked straight to that for Raspberry Pi Imager. It also has a handy video, which should explain the (fairly obvious) process in enough detail for most users.

I've tweaked the links to the 3 sub-pages as follows:

1. `NOOBS` - still linked, but added a note that it is now deprecated.
2 `Installing Images` - I've added a bit of blurb pointing folk to that page if they need some extra help, as in most cases raspberrypi.org/software/ should be all folk need. I've also pointed folk to the forums for extra help as well.
3. `SD cards` - I've labelled this "further info about SD cards", since in most cases the info on this page should not really be needed, but it's there if folk require it.