Restructure EEPROM documentation
Move text between the bootloader releases and re-order for clarity

msd.md
This page describes USB boot and the simplest methods for
enabling it.

bootloader_config
This is a technical reference for the bootloader configurtion file. Move
the tutorials for USB and bootloader updates to other pages.

boot_eeprom
The page documents that
1. There is an EEPROM
2. How to update it with the default configuration
3. How to change the high level bootmode
4. How to do custom configuration
5. Automatic updates
6. All the low level bits

Also:
* Documented the [config.txt] section of the bootloader EEPROM
* Add many anchors and some review comments