Update bcm2711_bootloader_config.md
The description of BCM-USB-MSD boot mode was not quite specific enough when it came to specifying which devices (Pi 4, CM4) have which type of USB connector. Also, the NVME boot description assumed there was a PCIe `socket`, which may not be the case on 3rd-party boards, so change that to `interface`.