Update net_tutorial.md (#1775)
Specifying Port=0 is not required and will actually cause problems if you are using this for other systems on the network needing DNS services. It should be omitted allowing the default port to be used. I assume this is probably originally put in here to prevent users from making a DNS server that is a security risk but there should at least be a comment about it as I have added here.