Update timelapse.md (#1932)
* Update timelapse.md

At line 54, removal of the underscore in 'image_*Jpg' for consistency with the creation of the images at line 10 (image%04d.jpg). This will avoid raising error below when people (e.g. me earlier today 😁) brutally copy-paste the example.
> [image2 @ 0x5579541a5780] Could not open file : frame-*.jpg
>[image2 @ 0x5579541a5780] Could not find codec parameters for stream 0 (Video: mjpeg, none(bt470bg/unknown/unknown)): unspecified size
>Consider increasing the value for the 'analyzeduration' and 'probesize' options

* Removal of underscore in lines 60-61 for consistency

As suggested by lurch, update also of the lines 61 and 62 for consistency with line 54.