Update docs to avoid rabbit-holing
Some very basic unstated assumptions exist for this document.  While the steps were easy to follow on Ubuntu, none of it worked trying to enable a screen and when I'd find guides on debugging they all seemed to assume the presence of `raspi-gpio` and `vcdbg`, which are apparently unavailable for Aarch64.