Append the extra .htaccess configuration
As redirects are processed in order, we need the most specific matches
to come first. As the redirects generated from the datasheets CSV are
specific URLs, they should therefore appear first in our .htaccess with
the "extra" configuration coming second (as this tends to be broader).

Without this, the redirects for the top-level documentation categories
break the more specific redirects for PDFs that now live at
https://datasheets.raspberrypi.org