Don't run build twice for pull requests
To prevent the GitHub Actions build running twice for pull requests
raised by contributors with write access to this repository, update our
configuration to only run the build for pull requests or pushes to the
master branch specifically. This way, pushes to other branches (e.g.
feature branches) will not trigger a build.