Update installing-from-an-image.adoc
Style and grammar fixups. Mostly removing extraneous `the`. Also Raspberry Pi style guide says not to use `the` in front of product names - I would assert that Raspberry Pi Imager is a product although it is a piece of software and is available for free.

I've also removed some duplicate which was created by the recent insertion of a paragraph advising that NOOBS is deprecated, and moved that paragraph out of the Raspberry Pi Imager section to a more appropriate location.

I've also removed two instances of scare quotes, which are not to be used, according to the Raspberry Pi style guide. I don't have access to the AP style guide, since it requires payment, which I have not made, but I would expect it also advises against using scare quotes.