Use head ref (not ref) for checking source branch
As ref will always be set to something like refs/pulls/123/merge for
pull requests, use head ref instead which reveals the source branch's
name.

As this would make it possible for someone to raise a PR from a fork
with the branch name "develop", we also add a check to ensure that
"master" can only be merged into from a branch local to this repository.