Fix text overflowing on configuration page
Google Search Console flagged that
https://www.raspberrypi.com/documentation/computers/configuration.html
has "content wider than screen". This appears to be caused by the line
beginning "There are a number of separate pins_* sections" which is
being rendered as a <pre> tag but with no scroll bars. Fix the markup so
it is rendered as a paragraph instead.