Make header consistent across raspberrypi.com
Brian noticed that the header changed sizes across Gatsby, WordPress and
the documentation site. This was due to a missing line-height reset for
the header and the introduction of significant whitespace in the header
links.

To fix the latter, we use the fact <li> elements do not need to be
closed in HTML5 as long as they are followed by another <li> or the last
item in the list. This avoids introducing any accidental whitespace. See
https://html.spec.whatwg.org/dev/syntax.html#optional-tags