Remove extra height from logo in header
While fixing the consistency of the header across raspberrypi.com, I
noticed an extra 1.5px of height added to the wrapper around the logo.
This seems to be coming from an inconsistent vertical alignment of the
logo in the header so explicitly set the alignment for both the logo and
its containing anchor tag.