Remove hdmi_enable_4kp60 (also in common.adoc)
==== `hdmi_enable_4kp60` (Raspberry Pi 4, Model B Only)

By default, when connected to a 4K monitor, the Raspberry Pi 4B will select a 30hz refresh rate. Use this option to allow selection of 60Hz refresh rates. Note, this will increase power consumption and increase the temperature of the Raspberry Pi. It is not possible to output 4Kp60 on both micro HDMI ports simultaneously.