Replace Simple Jekyll Search with DocSearch
In order to provide a better search experience and address some of the
usability issues raised in
https://github.com/raspberrypi/documentation/issues/1987, replace our
Simple Jekyll Search-powered search box with one provided by [Algolia
DocSearch][0].

We use the default, [out of the box theme provided by DocSearch v3][1]
as customising it will involve rewriting/overriding the CSS provided
directly by Algolia or overriding the default options in JS to render
our own markup. As it is, my hope is that it provides a better user
experience than the current search box.

  [0]: https://docsearch.algolia.com
  [1]: https://docsearch.algolia.com/docs/styling