Remove dependency on legacy stylesheet
The site currently loads a legacy stylesheet from the raspberrypi.com
WordPress theme because it was previously required to render the shared
site footer correctly. Since we switched to an entirely self-contained
footer with the launch of raspberrypi.com, there's no need to depend on
this stylesheet.

Remove it along with an unused web font from Google Fonts (again,
previously required by the old site footer) and add in any styles we
were implicitly inheriting from the old stylesheet, namely:

1. Links should be Raspberry Red and only have an underline on hover
2. Subheadings in the table of contents should have a line height of
   24px (1.5 times the base font size of 16px)

We add a timestamp when loading our stylesheets to bust any existing
browser caches otherwise anyone with cached versions of these styles
will not see the aforementioned updated styles.