Remove unused JSON search data
Now we're relying on Algolia DocSearch to host our search index, remove
the JSON index we were building for use with Simple Jekyll Search.

DocSearch is configured to crawl the site daily (at a random
time) and update its search index. This means there may be a 24 hour
delay before the index is updated unless we manually trigger a new crawl
from the Algolia Crawler dashboard.