Update installing-from-chromeos.adoc
As of the 2022-04-07 linked archive, the Raspberry Pi OS images are in XZ format and no longer in Zip format. There appears to be a bug in the Recovery Utility where it recognizes XZ files as image archives, but then the next screen blinks by quickly to the "success" screen without actually writing the image to the device. A workaround until the Recovery Utility is fixed is to first convert the XZ archive to Zip format by 'unxz'ing the original archive and then 'zip'ing a new archive. The Recovery Utility correctly writes the new zip archived image to the device. I tested this workaround with the raspios_full_armhf-2022-04-07 archive.