Bust cache for CSS and JS on every build
GitHub: https://github.com/raspberrypi/documentation/issues/2991

To prevent stale versions of our stylesheets and JavaScript from being
served after a deploy, ensure their URL is unique for each build by
appending the time of the build (as seconds since the epoch) in a query
string.

This means we will bust the cache even if our CSS and JS don't change
(as the query string changes on every build rather than being a function
of the contents of the relevant file) but this errs on the side of
caution and keeps the implementation very simple compared to introducing
an asset pipeline.

We only add the cache busting query string to files we change as opposed
to vendored JavaScript (jQuery, etc.) as they shouldn't change without
also updating their filename.