GS Camera external trigger: fix for #3237
Clarify that XTR is a 1.8V signal, and show a potential divier
with 1.5K and 1.8K resistors to generate it. Minor correction
that XTRIG would have been driven by GP1 (not GP0).

(Retain the instruction to remove R11 only if Q2 was fitted.)

Elsewhere, increase the suggested resistor values for the 1.65V
divider used to pull up XVS (2.2K was unnecessarily small).

Figures were hand-edited and may need to be regenerated by Fritzing.