Added warning about correct storage device when saying safe to write
The documentation tells people it is safe to proceed when they are using admin credentials to write to the target storage device. This is true, but only if they have followed the previous warning/instructions to be sure they have the correct storage device selected. When explicitly saying something is safe, it should be clear what the caveats to that are, as people often skip earlier sections.