Additions to audio HAT set up and configuration
Give specific mention to the presence of pwr indication when installing
a HAT. Provide some guidance on the usage of .soundrc and offer advice
on messages that can occur upon invocation of alsactl to load state
files. Removal of set -e will ensure alsactl won't cause the example
rc.local to exit prematurely with non-zero.