PARTITION parameters: Preliminary documentation for remapping partition numbers
The latest firmware provides a basic method for remapping a high
partition number to a regular partition. This makes it more flexible
as generic reboot parameter.
E.g. It would be possible reboot into a different boot-order or
even network install.
This will probably evolve as we use it with rpi-image-gen