Only respect system colour scheme preferences
Rather than implementing a JavaScript-based override on top of the
default user preferences for colour schemes, rely entirely on the user
preferences reported by the browser. This way, we can implement dark
mode support in pure CSS and remove the dependency on JavaScript
(without which, the site mistakenly would render in dark mode by
default).