Don't include subsections in the site-wide TOC
To reduce the number of elements included on every single page of the
documentation site, no longer include subsections of every section (e.g.
terminate at Pico C SDK > Hardware APIs > hardware_adc, rather than
also including the subsections of "Detailed Description", "Functions",
etc.) as these are available in the right-hand "On this page" TOC shown
on wide-enough screens.

Note we're not yet cutting down the amount of markup to render the TOC
but starting by cutting out the source data.

This removes the JavaScript logic to render the currently clicked
section in bold.