Upgrade to Tocbot 4.35.3 and simplify setup
Previously, the "On this page" table of contents would contain a
server-side rendered table of contents generated by the Jekyll AsciiDoc
plugin but this was always set to be hidden (even to screen readers). We
would then replace this element with the client-side rendered table of
contents generated by Tocbot. The latter is only visible on screens
wider than 1460px so there is no need for us to have logic to refresh
the Tocbot whenever the window is resized under 768px (as it will be
hidden anyway).