Continuously deploy to staging and production
Rather than deploying staging from one repository and production from
another, switch to using a single repository to deploy to both
environments.

We keep our pattern of a "develop" and "master" branch but merging to
the former will now deploy to staging and the latter to production.

We use [GitHub Actions Environments] [0] to vary the necessary secrets for
deployment and ensure only certain branches can deploy to a specific
environment.

  [0]: https://docs.github.com/en/actions/how-tos/deploy/configure-and-manage-deployments/manage-environments