Mirror primary repo to documentation repo
Set up a new GitHub Actions workflow to mirror the primary repository's
master branch to the public documentation repository. This requires the
use of a GitHub access token that has permission to read and write to
both repositories (and create workflows).

Note we do a full clone of the repository otherwise we can't force push
all changes to the documentation repository.

Now the documentation repository will only have a master branch, remove
any references to the develop branch.